requirements.txt
pillow
numpy
opencv-contrib-python
opencv-python
scipy
tqdm
